fix(src/inlineSVG.js): adds check for callback and parentNode
There was an odd issue where the callback was never getting called, because it was defined as a callback that didn't exist. This resolves that. This also adds a check for elements with the selectorClass that do not have a parentNode, or have already been converted. Also adds a check to see if a callback exists before calling one.